+++ core/src/ecs/cubos.cpp
@@ -2,6 +2,7 @@
 
 #include <cubos/core/ecs/command_buffer.hpp>
 #include <cubos/core/ecs/cubos.hpp>
+#include <cubos/core/ecs/name.hpp>
 #include <cubos/core/ecs/observer/observers.hpp>
 #include <cubos/core/log.hpp>
 #include <cubos/core/reflection/external/string.hpp>
@@ -9,35 +10,8 @@
 using cubos::core::ecs::Arguments;
 using cubos::core::ecs::Cubos;
 using cubos::core::ecs::DeltaTime;
+using cubos::core::ecs::Name;
 using cubos::core::ecs::ShouldQuit;
-using cubos::core::ecs::TagBuilder;
-
-TagBuilder::TagBuilder(World& world, core::ecs::Dispatcher& dispatcher, std::vector<std::string>& tags)
-    : mWorld(world)
-    , mDispatcher(dispatcher)
-    , mTags(tags)
-{
-}
-
-TagBuilder& TagBuilder::before(const Tag& tag)
-{
-    mTags.push_back(tag.id());
-    mDispatcher.tagSetBeforeTag(tag.id());
-    return *this;
-}
-
-TagBuilder& TagBuilder::after(const Tag& tag)
-{
-    mTags.push_back(tag.id());
-    mDispatcher.tagSetAfterTag(tag.id());
-    return *this;
-}
-
-TagBuilder& TagBuilder::tagged(const Tag& tag)
-{
-    mDispatcher.tagInheritTag(tag.id());
-    return *this;
-}
 
 Cubos::Cubos()
     : Cubos(1, nullptr)
@@ -48,17 +22,111 @@ Cubos::Cubos(int argc, char** argv)
 {
     std::vector<std::string> arguments(argv + 1, argv + argc);
 
-    this->addResource<DeltaTime>();
-    this->addResource<ShouldQuit>();
-    this->addResource<Arguments>(Arguments{.value = arguments});
+    this->resource<DeltaTime>();
+    this->resource<ShouldQuit>();
+    this->resource<Arguments>(Arguments{.value = arguments});
+}
+
+Cubos& Cubos::plugin(Plugin plugin)
+{
+    CUBOS_ASSERT(!mInstalledPlugins.contains(plugin), "Plugin has already been added by another plugin");
+
+    mInstalledPlugins.at(mPluginStack.back()).subPlugins.emplace(plugin);
+    mInstalledPlugins.insert({plugin, PluginInfo{}});
+
+    mPluginStack.push_back(plugin);
+    plugin(*this);
+    mPluginStack.pop_back();
+
+    return *this;
+}
+
+Cubos& Cubos::depends(Plugin plugin)
+{
+    CUBOS_ASSERT(mInstalledPlugins.contains(plugin),
+                 "Plugin dependency wasn't fulfilled. Did you forget to add a plugin?");
+
+    mInstalledPlugins.at(plugin).dependentCount += 1;
+    mInstalledPlugins.at(mPluginStack.back()).dependencies.emplace(plugin);
+
+    return *this;
+}
+
+Cubos& Cubos::component(const reflection::Type& type)
+{
+    CUBOS_ASSERT(!mTypeToPlugin.contains(type), "Component {} was already registered", type.name());
+    mTypeToPlugin.insert(type, mPluginStack.back());
+    mWorld.registerComponent(type);
+    return *this;
+}
+
+Cubos& Cubos::relation(const reflection::Type& type)
+{
+    CUBOS_ASSERT(!mTypeToPlugin.contains(type), "Relation {} was already registered", type.name());
+    mTypeToPlugin.insert(type, mPluginStack.back());
+    mWorld.registerRelation(type);
+    return *this;
+}
+
+Cubos::TagBuilder Cubos::tag(const Tag& tag)
+{
+    if (!this->isRegistered(tag))
+    {
+        CUBOS_ASSERT(!mTagToPlugin.contains(tag.id()),
+                     "Tag {} was already registered by another plugin - if you depend on it, define the dependency",
+                     tag.name());
+        mTagToPlugin.emplace(tag.id(), mPluginStack.back());
+    }
+
+    mMainDispatcher.addTag(tag.id());
+    return TagBuilder{*this, mMainDispatcher};
+}
+
+Cubos::TagBuilder Cubos::startupTag(const Tag& tag)
+{
+    if (!this->isRegistered(tag))
+    {
+        CUBOS_ASSERT(!mTagToPlugin.contains(tag.id()),
+                     "Tag {} was already registered by another plugin - if you depend on it, define the dependency",
+                     tag.name());
+        mTagToPlugin.emplace(tag.id(), mPluginStack.back());
+    }
+
+    mStartupDispatcher.addTag(tag.id());
+    return TagBuilder{*this, mStartupDispatcher};
+}
+
+Cubos::TagBuilder Cubos::noTag(const Tag& tag)
+{
+    CUBOS_ASSERT(this->isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
+    mMainDispatcher.addNegativeTag(tag.id());
+    return TagBuilder{*this, mMainDispatcher};
+}
+
+Cubos::TagBuilder Cubos::noStartupTag(const Tag& tag)
+{
+    CUBOS_ASSERT(this->isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
+    mStartupDispatcher.addNegativeTag(tag.id());
+    return TagBuilder{*this, mStartupDispatcher};
+}
+
+auto Cubos::system(std::string name) -> SystemBuilder
+{
+    return {*this, mMainDispatcher, std::move(name)};
+}
+
+auto Cubos::startupSystem(std::string name) -> SystemBuilder
+{
+    return {*this, mStartupDispatcher, std::move(name)};
+}
+
+auto Cubos::observer(std::string name) -> ObserverBuilder
+{
+    return {*this, std::move(name)};
 }
 
 void Cubos::run()
 {
-    mPlugins.clear();
-    mMainTags.clear();
-    mStartupTags.clear();
-
     // Compile execution chain
     mStartupDispatcher.compileChain();
     mMainDispatcher.compileChain();
@@ -80,61 +148,77 @@ void Cubos::run()
     } while (!mWorld.read<ShouldQuit>().get().value);
 }
 
-Cubos& Cubos::addPlugin(void (*func)(Cubos&))
+bool Cubos::isRegistered(const reflection::Type& type) const
 {
-    if (!mPlugins.contains(func))
+    return (mTypeToPlugin.contains(type) && this->isKnownPlugin(mTypeToPlugin.at(type), mPluginStack.back())) ||
+           type.is<Name>();
+}
+
+bool Cubos::isRegistered(const Tag& tag) const
+{
+    return mTagToPlugin.contains(tag.id()) && this->isKnownPlugin(mTagToPlugin.at(tag.id()), mPluginStack.back());
+}
+
+bool Cubos::isSubPlugin(Plugin subPlugin, Plugin basePlugin) const
+{
+    for (auto p : mInstalledPlugins.at(basePlugin).subPlugins)
     {
-        func(*this);
-        mPlugins.insert(func);
+        if (p == subPlugin || this->isSubPlugin(subPlugin, p))
+        {
+            return true;
+        }
     }
-    else
+
+    return false;
+}
+
+bool Cubos::isKnownPlugin(Plugin plugin, Plugin basePlugin) const
+{
+    if (plugin == basePlugin || this->isSubPlugin(plugin, basePlugin))
     {
-        CUBOS_TRACE("Plugin was already registered!");
+        return true;
     }
+
+    for (auto dependency : mInstalledPlugins.at(basePlugin).dependencies)
+    {
+        if (plugin == dependency || this->isSubPlugin(plugin, dependency))
+        {
+            return true;
+        }
+    }
+
+    return false;
+}
+
+Cubos::TagBuilder::TagBuilder(Cubos& cubos, core::ecs::Dispatcher& dispatcher)
+    : mCubos(cubos)
+    , mDispatcher(dispatcher)
+{
+}
+
+Cubos::TagBuilder& Cubos::TagBuilder::before(const Tag& tag)
+{
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
+    mDispatcher.tagSetBeforeTag(tag.id());
     return *this;
 }
 
-TagBuilder Cubos::tag(const Tag& tag)
+Cubos::TagBuilder& Cubos::TagBuilder::after(const Tag& tag)
 {
-    mMainDispatcher.addTag(tag.id());
-    return TagBuilder{mWorld, mMainDispatcher, mMainTags};
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
+    mDispatcher.tagSetAfterTag(tag.id());
+    return *this;
 }
 
-TagBuilder Cubos::startupTag(const Tag& tag)
+Cubos::TagBuilder& Cubos::TagBuilder::tagged(const Tag& tag)
 {
-    mStartupDispatcher.addTag(tag.id());
-    return TagBuilder{mWorld, mStartupDispatcher, mStartupTags};
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
+    mDispatcher.tagInheritTag(tag.id());
+    return *this;
 }
 
-TagBuilder Cubos::noTag(const Tag& tag)
-{
-    mMainDispatcher.addNegativeTag(tag.id());
-    return TagBuilder{mWorld, mMainDispatcher, mMainTags};
-}
-
-TagBuilder Cubos::noStartupTag(const Tag& tag)
-{
-    mStartupDispatcher.addNegativeTag(tag.id());
-    return TagBuilder{mWorld, mStartupDispatcher, mStartupTags};
-}
-
-auto Cubos::system(std::string name) -> SystemBuilder
-{
-    return {mWorld, mMainDispatcher, std::move(name)};
-}
-
-auto Cubos::startupSystem(std::string name) -> SystemBuilder
-{
-    return {mWorld, mStartupDispatcher, std::move(name)};
-}
-
-auto Cubos::observer(std::string name) -> ObserverBuilder
-{
-    return {mWorld, std::move(name)};
-}
-
-Cubos::SystemBuilder::SystemBuilder(World& world, Dispatcher& dispatcher, std::string name)
-    : mWorld{world}
+Cubos::SystemBuilder::SystemBuilder(Cubos& cubos, Dispatcher& dispatcher, std::string name)
+    : mCubos{cubos}
     , mDispatcher{dispatcher}
     , mName{std::move(name)}
 {
@@ -142,18 +226,21 @@ Cubos::SystemBuilder::SystemBuilder(World& world, Dispatcher& dispatcher, std::s
 
 auto Cubos::SystemBuilder::tagged(const Tag& tag) && -> SystemBuilder&&
 {
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
     mTagged.insert(tag.id());
     return std::move(*this);
 }
 
 auto Cubos::SystemBuilder::before(const Tag& tag) && -> SystemBuilder&&
 {
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
     mBefore.insert(tag.id());
     return std::move(*this);
 }
 
 auto Cubos::SystemBuilder::after(const Tag& tag) && -> SystemBuilder&&
 {
+    CUBOS_ASSERT(mCubos.isRegistered(tag), "Tag {} wasn't registered by the plugin or its dependencies", tag.name());
     mAfter.insert(tag.id());
     return std::move(*this);
 }
@@ -180,9 +267,9 @@ auto Cubos::SystemBuilder::entity(int target) && -> SystemBuilder&&
 
 auto Cubos::SystemBuilder::with(const reflection::Type& type, int target) && -> SystemBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
 
     if (mOptions.empty())
     {
@@ -204,9 +291,9 @@ auto Cubos::SystemBuilder::with(const reflection::Type& type, int target) && -> 
 
 auto Cubos::SystemBuilder::without(const reflection::Type& type, int target) && -> SystemBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
 
     if (mOptions.empty())
     {
@@ -234,11 +321,11 @@ auto Cubos::SystemBuilder::related(const reflection::Type& type, int fromTarget,
 auto Cubos::SystemBuilder::related(const reflection::Type& type, Traversal traversal, int fromTarget,
                                    int toTarget) && -> SystemBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such relation type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isRelation(dataTypeId), "Type {} isn't registered as a relation", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such relation type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isRelation(dataTypeId), "Type {} isn't registered as a relation", type.name());
 
-    CUBOS_ASSERT_IMP(traversal != Traversal::Random, mWorld.types().isTreeRelation(dataTypeId),
+    CUBOS_ASSERT_IMP(traversal != Traversal::Random, mCubos.mWorld.types().isTreeRelation(dataTypeId),
                      "Directed traversal can only be used for tree relations, and {} isn't registered as one",
                      type.name());
 
@@ -271,6 +358,23 @@ auto Cubos::SystemBuilder::other() && -> SystemBuilder&&
 
 void Cubos::SystemBuilder::finish(System<void> system)
 {
+    if (mCondition.contains())
+    {
+        for (const auto& dataTypeId : mCondition.value().access().dataTypes)
+        {
+            const auto& type = mCubos.mWorld.types().type(dataTypeId);
+            CUBOS_ASSERT(mCubos.isRegistered(type), "Type {} isn't registered by the plugin or its dependencies",
+                         type.name());
+        }
+    }
+
+    for (const auto& dataTypeId : system.access().dataTypes)
+    {
+        const auto& type = mCubos.mWorld.types().type(dataTypeId);
+        CUBOS_ASSERT(mCubos.isRegistered(type), "Type {} isn't registered by the plugin or its dependencies",
+                     type.name());
+    }
+
     CUBOS_DEBUG("Added system {}", mName);
     mDispatcher.addSystem(std::move(mName), std::move(system));
 
@@ -295,17 +399,17 @@ void Cubos::SystemBuilder::finish(System<void> system)
     }
 }
 
-Cubos::ObserverBuilder::ObserverBuilder(World& world, std::string name)
-    : mWorld{world}
+Cubos::ObserverBuilder::ObserverBuilder(Cubos& cubos, std::string name)
+    : mCubos{cubos}
     , mName{std::move(name)}
 {
 }
 
 auto Cubos::ObserverBuilder::onAdd(const reflection::Type& type, int target) && -> ObserverBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
     CUBOS_ASSERT(mColumnId == ColumnId::Invalid, "An observer can only have at most one hook");
 
     if (mOptions.empty())
@@ -324,15 +428,15 @@ auto Cubos::ObserverBuilder::onAdd(const reflection::Type& type, int target) && 
 
     mOptions.back().observedTarget = target;
     mRemove = false;
-    mColumnId = ColumnId::make(mWorld.types().id(type));
+    mColumnId = ColumnId::make(mCubos.mWorld.types().id(type));
     return std::move(*this);
 }
 
 auto Cubos::ObserverBuilder::onRemove(const reflection::Type& type, int target) && -> ObserverBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
     CUBOS_ASSERT(mColumnId == ColumnId::Invalid, "An observer can only have at most one hook");
 
     if (mOptions.empty())
@@ -351,7 +455,7 @@ auto Cubos::ObserverBuilder::onRemove(const reflection::Type& type, int target) 
 
     mOptions.back().observedTarget = target;
     mRemove = true;
-    mColumnId = ColumnId::make(mWorld.types().id(type));
+    mColumnId = ColumnId::make(mCubos.mWorld.types().id(type));
     return std::move(*this);
 }
 
@@ -377,9 +481,9 @@ auto Cubos::ObserverBuilder::entity(int target) && -> ObserverBuilder&&
 
 auto Cubos::ObserverBuilder::with(const reflection::Type& type, int target) && -> ObserverBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
 
     if (mOptions.empty())
     {
@@ -401,9 +505,9 @@ auto Cubos::ObserverBuilder::with(const reflection::Type& type, int target) && -
 
 auto Cubos::ObserverBuilder::without(const reflection::Type& type, int target) && -> ObserverBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such component type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such component type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isComponent(dataTypeId), "Type {} isn't registered as a component", type.name());
 
     if (mOptions.empty())
     {
@@ -431,11 +535,11 @@ auto Cubos::ObserverBuilder::related(const reflection::Type& type, int fromTarge
 auto Cubos::ObserverBuilder::related(const reflection::Type& type, Traversal traversal, int fromTarget,
                                      int toTarget) && -> ObserverBuilder&&
 {
-    CUBOS_ASSERT(mWorld.types().contains(type), "No such relation type {} was registered", type.name());
-    auto dataTypeId = mWorld.types().id(type);
-    CUBOS_ASSERT(mWorld.types().isRelation(dataTypeId), "Type {} isn't registered as a relation", type.name());
+    CUBOS_ASSERT(mCubos.isRegistered(type), "No such relation type {} was registered", type.name());
+    auto dataTypeId = mCubos.mWorld.types().id(type);
+    CUBOS_ASSERT(mCubos.mWorld.types().isRelation(dataTypeId), "Type {} isn't registered as a relation", type.name());
 
-    CUBOS_ASSERT_IMP(traversal != Traversal::Random, mWorld.types().isTreeRelation(dataTypeId),
+    CUBOS_ASSERT_IMP(traversal != Traversal::Random, mCubos.mWorld.types().isTreeRelation(dataTypeId),
                      "Directed traversal can only be used for tree relations, and {} isn't registered as one",
                      type.name());
 
@@ -468,15 +572,22 @@ auto Cubos::ObserverBuilder::other() && -> ObserverBuilder&&
 
 void Cubos::ObserverBuilder::finish(System<void> system)
 {
+    for (const auto& dataTypeId : system.access().dataTypes)
+    {
+        const auto& type = mCubos.mWorld.types().type(dataTypeId);
+        CUBOS_ASSERT(mCubos.isRegistered(type), "Type {} isn't registered by the plugin or its dependencies",
+                     type.name());
+    }
+
     CUBOS_ASSERT(mColumnId != ColumnId::Invalid, "You must set at least one hook for the observer");
     CUBOS_DEBUG("Added observer {}", mName);
 
     if (mRemove)
     {
-        mWorld.observers().hookOnRemove(mColumnId, std::move(system));
+        mCubos.mWorld.observers().hookOnRemove(mColumnId, std::move(system));
     }
     else
     {
-        mWorld.observers().hookOnAdd(mColumnId, std::move(system));
+        mCubos.mWorld.observers().hookOnAdd(mColumnId, std::move(system));
     }
 }