+++ engine/samples/assets/bridge/main.cpp
@@ -52,7 +52,8 @@ int main()
 {
     auto cubos = Cubos();
 
-    cubos.addPlugin(assetsPlugin);
+    cubos.plugin(settingsPlugin);
+    cubos.plugin(assetsPlugin);
 
     cubos.startupSystem("configure Assets plugin").tagged(settingsTag).call([](Settings& settings) {
         settings.setString("assets.io.path", SAMPLE_ASSETS_FOLDER);