+++ CHANGELOG.md
@@ -24,10 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rotated box Gizmo (#878, **@DiogoMendonc-a**).
 - Shader asset and bridge (#1058, **@tomas7770**).
 - GBuffer plugin (#1061, **@RiscadoA**).
+- Defaults plugin (#229, **@RiscadoA**).
 
 ### Changed
 
 - Make reflection work for all types, even those without reflection implemented (#1092, **@RiscadoA**).
+- Force plugins to be added only once (#422, **@RiscadoA**).
 
 ### Fixed